requirements.txt

# RPi interfacing
pigpio

# sensors
tfluna-driver
w1thermsensor

# monitoring
prometheus_client